id generation is reworked, all steps of concert creation via EventManagementController are created

+++ src/main/java/com/hantel/event_manager/service/MusicalService.java
@@ -16,7 +16,7 @@ public class MusicalService {
         this.musicalRepository = musicalRepository;
     }
 
-    public Musical findById(int id) {
+    public Musical findById(Long id) {
         return musicalRepository.findById(id);
     }
